Refactor, move SpriteHandlers to own package
