+++ src/main/java/inf112/roborally/app/main/MapLayout.java
@@ -9,7 +9,7 @@ import inf112.roborally.app.editor.Console;
 import inf112.roborally.app.editor.EditorInput;
 import inf112.roborally.app.exceptions.OutsideGridException;
 import inf112.roborally.app.player.Player;
-import inf112.roborally.app.tile.SpriteLoader;
+import inf112.roborally.app.sprite.SpriteLoader;
 
 /**
  * Class used to render the Maps layout
