+++ src/main/java/inf112/roborally/app/tile/tiles/AbstractTile.java
@@ -2,7 +2,7 @@ package inf112.roborally.app.tile.tiles;
 
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import inf112.roborally.app.tile.IBoardTile;
-import inf112.roborally.app.tile.SpriteLoader;
+import inf112.roborally.app.sprite.SpriteLoader;
 
 public abstract class AbstractTile implements IBoardTile {
